auth

+++ api/users/users.controller.js
@@ -6,14 +6,6 @@ const usersService = require('./users.service')
 const articleService = require('../articles/articles.service')
 
 class UsersController {
-  async getAll(req, res, next) {
-    try {
-      const users = await usersService.getAll()
-      res.json(users)
-    } catch (err) {
-      next(err)
-    }
-  }
   async getById(req, res, next) {
     try {
       const id = req.params.id
@@ -26,8 +18,14 @@ class UsersController {
       next(err)
     }
   }
+
   async create(req, res, next) {
+    const authUser = req.user
     try {
+      if (authUser.role !== 'admin') {
+        throw new UnauthorizedError()
+      }
+
       const user = await usersService.create(req.body)
       user.password = undefined
       req.io.emit('user:create', user)
@@ -36,10 +34,17 @@ class UsersController {
       next(err)
     }
   }
+
   async update(req, res, next) {
+    const user = req.user
     try {
       const id = req.params.id
       const data = req.body
+
+      if (user.role !== 'admin' && user._id.toString() !== id) {
+        throw new UnauthorizedError()
+      }
+
       const userModified = await usersService.update(id, data)
       userModified.password = undefined
       res.json(userModified)
@@ -47,9 +52,15 @@ class UsersController {
       next(err)
     }
   }
+
   async delete(req, res, next) {
+    const user = req.user
     try {
       const id = req.params.id
+      if (user.role !== 'admin' && user._id.toString() !== id) {
+        throw new UnauthorizedError()
+      }
+
       await usersService.delete(id)
       req.io.emit('user:delete', { id })
       res.status(204).send()
@@ -57,6 +68,7 @@ class UsersController {
       next(err)
     }
   }
+
   async login(req, res, next) {
     try {
       const { email, password } = req.body
@@ -64,6 +76,7 @@ class UsersController {
       if (!userId) {
         throw new UnauthorizedError()
       }
+
       const token = jwt.sign({ userId }, config.secretJwtToken, {
         expiresIn: '3d',
       })
@@ -77,11 +90,16 @@ class UsersController {
   }
 
   async getUser(req, res, next) {
+    const authUser = req.user
     try {
       const decoded = jwt.verify(req.params.id, config.secretJwtToken)
       const user = await usersService.get(decoded.userId)
       if (!user) {
         throw new NotFoundError()
+      }
+
+      if (authUser.role !== 'admin' && authUser._id.toString() !== user._id.toString()) {
+        throw new UnauthorizedError()
       }
 
       res.json(user)
@@ -91,11 +109,30 @@ class UsersController {
   }
 
   async articles(req, res, next) {
+    const user = req.user
     try {
       const userId = req.params.id
-      const articles = await articleService.getArticlesByUser(userId)
 
+      if (user.role !== 'admin' && user._id.toString() !== userId) {
+        throw new UnauthorizedError()
+      }
+
+      const articles = await articleService.getArticlesByUser(userId)
       res.status(200).json(articles)
+    } catch (err) {
+      next(err)
+    }
+  }
+
+  async getAll(req, res, next) {
+    const user = req.user
+    try {
+      if (user.role !== 'admin') {
+        throw new UnauthorizedError()
+      }
+
+      const users = await usersService.getAll()
+      res.json(users)
     } catch (err) {
       next(err)
     }
